Drops extra queue for InvalidationProcessor, merges completion Runnable and threshold changed callbacks into ScratchoffController.Delegate,  updates tests/samples/readme, marks version as 2.0.0, adds upgrade migration guide

+++ scratchoff/src/androidTest/java/com/jackpocket/scratchoff/processors/ScratchoffProcessorTests.kt
@@ -20,7 +20,7 @@ class ScratchoffProcessorTests {
                     collectedPaths.addAll(it)
                 },
                 object: ThresholdProcessor(0, 0.0, object: ThresholdProcessor.Delegate {
-                    override fun postScratchPercentChanged(percent: Double) { }
+                    override fun postScratchPercentChanged(percent: Float) { }
                     override fun postScratchThresholdReached() { }
                     override fun getScratchableLayoutSize(): IntArray {
                         return intArrayOf()
@@ -67,7 +67,7 @@ class ScratchoffProcessorTests {
         val processor = ScratchoffProcessor(
                 null,
                 object: ThresholdProcessor(0, 0.0, object: ThresholdProcessor.Delegate {
-                    override fun postScratchPercentChanged(percent: Double) { }
+                    override fun postScratchPercentChanged(percent: Float) { }
                     override fun postScratchThresholdReached() { }
                     override fun getScratchableLayoutSize(): IntArray {
                         return intArrayOf()
